Moved the last dllimport from CORE to UI
