+++ src/LogExpert.Core/Config/ColorMode.cs
@@ -74,34 +74,5 @@ namespace LogExpert.Core.Config
             InactiveTabColor = LessBrightBackgroundColor;
             DarkModeEnabled = false;
         }
-
-        #region TitleBarDarkMode
-        [DllImport("dwmapi.dll")]
-        private static extern int DwmSetWindowAttribute(nint hwnd, int attr, ref int attrValue, int attrSize);
-
-        private const int DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_20H1 = 19;
-        private const int DWMWA_USE_IMMERSIVE_DARK_MODE = 20;
-
-        public static bool UseImmersiveDarkMode(nint handle, bool enabled)
-        {
-
-            var attribute = DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_20H1;
-            if (IsWindows10OrGreater(18985))
-            {
-                attribute = DWMWA_USE_IMMERSIVE_DARK_MODE;
-            }
-
-            int useImmersiveDarkMode = enabled ? 1 : 0;
-            return DwmSetWindowAttribute(handle, attribute, ref useImmersiveDarkMode, sizeof(int)) == 0;
-
-        }
-
-        private static bool IsWindows10OrGreater(int build = -1)
-        {
-            return Environment.OSVersion.Version.Major >= 10 && Environment.OSVersion.Version.Build >= build;
-        }
-
-        #endregion TitleBarDarkMode
-
     }
 }
