+++ src/LogExpert.UI/Extensions/Win32.cs
@@ -2,10 +2,10 @@
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 
-namespace LogExpert.Core.Classes
+namespace LogExpert.UI.Extensions
 {
     [SupportedOSPlatform("windows")]
-    public static class Win32
+    internal static class Win32
     {
         #region Fields
 
@@ -114,6 +114,34 @@ namespace LogExpert.Core.Classes
             ref nint iconsSmall,
             uint numIcons
         );
+
+        #region TitleBarDarkMode
+        [DllImport("dwmapi.dll")]
+        private static extern int DwmSetWindowAttribute(nint hwnd, int attr, ref int attrValue, int attrSize);
+
+        private const int DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_20H1 = 19;
+        private const int DWMWA_USE_IMMERSIVE_DARK_MODE = 20;
+
+        public static bool UseImmersiveDarkMode(nint handle, bool enabled)
+        {
+
+            var attribute = DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_20H1;
+            if (IsWindows10OrGreater(18985))
+            {
+                attribute = DWMWA_USE_IMMERSIVE_DARK_MODE;
+            }
+
+            int useImmersiveDarkMode = enabled ? 1 : 0;
+            return DwmSetWindowAttribute(handle, attribute, ref useImmersiveDarkMode, sizeof(int)) == 0;
+
+        }
+
+        private static bool IsWindows10OrGreater(int build = -1)
+        {
+            return Environment.OSVersion.Version.Major >= 10 && Environment.OSVersion.Version.Build >= build;
+        }
+
+        #endregion TitleBarDarkMode
 
         #endregion
     }
